requirements.txt.txt
streamlit
python-docx
pillow
pytesseract
PyMuPDF
google-generativeai